Simplify tests suite
* No need for an integration test without remote,
  moving all to test_binary_sensor
* Just 1 hass fixture with 2 bridges
