+++ tests/test_device_tracker.py
@@ -2,6 +2,7 @@
 from datetime import timedelta
 from unittest.mock import MagicMock, patch
 
+from custom_components.huesensor import DOMAIN
 from custom_components.huesensor.device_tracker import (
     async_setup_scanner,
     HueDeviceScanner,
@@ -10,7 +11,7 @@ from custom_components.huesensor.device_tracker import (
 from .conftest import MockAsyncCounter
 
 
-async def test_device_tracker_setup(mock_hass_2_bridges):
+async def test_device_tracker_setup(mock_hass):
     """Test platform setup for binary sensors."""
     mock_async_see = MockAsyncCounter()
     with patch(
@@ -18,8 +19,8 @@ async def test_device_tracker_setup(mock_hass_2_bridges):
         autospec=True,
     ) as mock_track_time:
         await async_setup_scanner(
-            mock_hass_2_bridges,
-            {"platform": "huesensor", "scan_interval": timedelta(seconds=10)},
+            mock_hass,
+            {"platform": DOMAIN, "scan_interval": timedelta(seconds=10)},
             mock_async_see,
         )
 
@@ -27,7 +28,7 @@ async def test_device_tracker_setup(mock_hass_2_bridges):
         assert mock_async_see.call_count == 1
 
 
-async def test_device_scanner(mock_hass, mock_hass_2_bridges):
+async def test_device_scanner(mock_hass):
     """Test platform setup for binary sensors."""
     mock_async_see = MockAsyncCounter()
 
@@ -36,12 +37,7 @@ async def test_device_scanner(mock_hass, mock_hass_2_bridges):
     await empty_scanner.async_update_info()
     assert mock_async_see.call_count == 0
 
-    # sensors, but no geofence data
-    scanner = HueDeviceScanner(mock_hass, mock_async_see)
-    await scanner.async_update_info()
-    assert mock_async_see.call_count == 0
-
     # 1 geofence in 2nd bridge
-    scanner = HueDeviceScanner(mock_hass_2_bridges, mock_async_see)
+    scanner = HueDeviceScanner(mock_hass, mock_async_see)
     await scanner.async_update_info()
     assert mock_async_see.call_count == 1
